+++ lilith/src/main/java/de/huxhorn/lilith/services/clipboard/LoggingCallStackFormatter.java
@@ -1,6 +1,6 @@
 /*
  * Lilith - a log event viewer.
- * Copyright (C) 2007-2011 Joern Huxhorn
+ * Copyright (C) 2007-2016 Joern Huxhorn
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -96,5 +96,10 @@ public class LoggingCallStackFormatter
 		}
 
 		return null;
+	}
+
+	public boolean isNative()
+	{
+		return true;
 	}
 }