+++ lilith/src/main/java/de/huxhorn/lilith/services/clipboard/LoggingMarkerFormatter.java
@@ -1,6 +1,6 @@
 /*
  * Lilith - a log event viewer.
- * Copyright (C) 2007-2011 Joern Huxhorn
+ * Copyright (C) 2007-2016 Joern Huxhorn
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -86,6 +86,11 @@ public class LoggingMarkerFormatter
 		}
 
 		return null;
+	}
+
+	public boolean isNative()
+	{
+		return true;
 	}
 
 	private void buildMarker(StringBuilder text, int indent, Marker marker, List<String> handledMarkers)